Fix_authorized User

+++ src/components/SectionSup/ModalEditCard/index.tsx
@@ -10,10 +10,13 @@ import { AuthContext } from '@/context/auth.context';
 
 export const ModalEditCard = ({ infoCard, setOpenModalEdit }:{ infoCard: iCardSup, setOpenModalEdit: Dispatch<SetStateAction<boolean>> }) => {
 
-    const { allUser } = useContext(AuthContext);
+    const { allUser, userId } = useContext(AuthContext);
     const { excluirSupCard, editarCard, excluirTask } = useContext(SupContext);
 
-    
+    const isWorker = infoCard.workers && infoCard.workers.some((worker: any) => worker.id === userId);
+    const isOwner = infoCard.userId === userId;
+    const authorized = isWorker || isOwner;
+
     const tasksinfoCard: any = infoCard.tasks!
     const [tasksDB, setTasksDB] = useState<any>(tasksinfoCard);
     const [novaTarefa, setNovaTarefa] = useState('');
@@ -34,8 +37,10 @@ export const ModalEditCard = ({ infoCard, setOpenModalEdit }:{ infoCard: iCardSu
 
     const [confirmacaoExclusao, setConfirmacaoExclusao] = useState(false);
     const confirmaExcluir = (cardId: string)=> {
-        excluirSupCard(cardId)
-        setOpenModalEdit(false)
+        if (authorized) {
+            excluirSupCard(cardId)
+            setOpenModalEdit(false)
+        }
     }
 
     const excluirTarefa = (index: number, id: any | undefined) => {
@@ -49,7 +54,6 @@ export const ModalEditCard = ({ infoCard, setOpenModalEdit }:{ infoCard: iCardSu
     
     const [workersApi, setWorkersApi] = useState<any | []>(infoCard.workers! || []);
     const [selectedUserIds, setSelectedUserIds] = useState<any>([]);
-    
     const handleUserChange = (event: React.ChangeEvent<HTMLSelectElement>) => {
         const userId = event.target.value;
         if (userId !== '') {
@@ -63,13 +67,11 @@ export const ModalEditCard = ({ infoCard, setOpenModalEdit }:{ infoCard: iCardSu
 
     const handleRemoveUser = (userIdToRemove: string) => {
         const updatedSelectedUserIds = selectedUserIds.filter((id: any) => id !== userIdToRemove);
-        console.log(updatedSelectedUserIds)
         setSelectedUserIds(updatedSelectedUserIds);
     };
 
     const handleRemoveUserAPI = (userIdToRemove: string) => {
         const updatedWorkersApi: any = workersApi.filter((user: any) => user.id !== userIdToRemove);
-        console.log(updatedWorkersApi)
         setWorkersApi(updatedWorkersApi);
     }
 
@@ -90,7 +92,7 @@ export const ModalEditCard = ({ infoCard, setOpenModalEdit }:{ infoCard: iCardSu
         <section className={styled.modal}>
             <div className={styled.modalCard}>
                 <form className={styled.form} onSubmit={handleSubmit(onSubmit)}>
-                     <div className={styled.divHeader}>
+                    <div className={styled.divHeader}>
                         <h1 className={styled.title}>{infoCard.title}</h1>
                         <button className={styled.btnFecha} type='button' onClick={() => setOpenModalEdit(false)}>Fechar</button>
                     </div>
@@ -98,51 +100,60 @@ export const ModalEditCard = ({ infoCard, setOpenModalEdit }:{ infoCard: iCardSu
                         <h3 className={styled.h3Name}>{infoCard.user?.name}</h3>
                         <h3 className={styled.h3Data}>{infoCard.createdAt}</h3>
                     </div>
-
+                    {!authorized && (
+                        <p className={styled.pAuthorized}>Você não é autorizado a fazer edições</p>
+                    )}
                     <div className={styled.divSelect}>
                         <div className={styled.divPriority}>
                             <p className={styled.pDesc}>Prioridade: </p>
-                            <select className={styled.select} id="opcoes" {...register('priority')} defaultValue={infoCard.priority || 'Normal'}>
+                            <select className={styled.select} id="opcoes" {...register('priority')} defaultValue={infoCard.priority || 'Normal'} disabled={!authorized}>
                                 <option value="Basica">Basica</option>
                                 <option value="Normal">Normal</option>
                                 <option value="Urgente">Urgente</option>
                                 <option value="Muito Urgente">Muito Urgente</option>
                             </select>
                         </div>
-
+    
                         <div className={styled.divWorkes}>
                             <div className={styled.divWorkesUser}>
                                 <p>Usuário:</p>
                                 <div className={styled.divUserSelect}>
-                                    <select className={styled.select} onBlur={handleUserChange}>
-                                        <option value=''>-</option>
-                                        {allUser && allUser.map((user) => (
-                                            <option key={user.id} value={user.id}>{user.name}</option>
-                                            ))}
+                                    <select className={styled.select} onBlur={handleUserChange} disabled={!authorized}>
+                                        <option value=''> - Selecionar - </option>
+                                        {allUser && allUser
+                                            .filter(user => user.id !== infoCard.userId)
+                                            .map(user => (
+                                                <option key={user.id} value={user.id}>{user.name}</option>
+                                            ))
+                                        }
                                     </select>
-                                    <button type='button'>Add</button>
+                                    <button type='button' disabled={!authorized}>Add</button>
                                 </div>
                             </div>
                             <div className={styled.divWorkesUserAPI}>
+                                <p>Autorizados:</p>
+                                <div key={userId} className={styled.divNameWorkes}>
+                                    <p>{infoCard.user.name}</p>
+                                </div>
                                 {workersApi.map((user: any) => (
                                     <div key={user.id} className={styled.divNameWorkes}>
                                         <p>{user.name}</p>
-                                        <button type='button' onClick={() => handleRemoveUserAPI(user.id)}>X</button>
+                                        <button type='button' onClick={() => handleRemoveUserAPI(user.id)} disabled={!authorized}>X</button>
                                     </div>
                                 ))}
                                 {selectedUserIds.map((userId: any) => (
                                     <div key={userId} className={styled.divNameWorkes}>
                                         {allUser!.find((user: any) => user.id === userId)?.name}
-                                        <button type='button' onClick={() => handleRemoveUser(userId)}>X</button>
+                                        <button type='button' onClick={() => handleRemoveUser(userId)} disabled={!authorized}>X</button>
                                     </div>
                                 ))}
                             </div>
                         </div>
                     </div>
-
+    
                     <div className={styled.divDesc}>
                         <p className={styled.pDesc}>Descrição</p>
-                        <textarea className={styled.textarea} id="descriptin" {...register("description")} defaultValue={infoCard.description || ''}></textarea>
+                        <textarea className={styled.textarea} id="descriptin" {...register("description")} defaultValue={infoCard.description || ''} readOnly={!authorized}></textarea>
                     </div>
                     <div className={styled.divTarefas}>
                         <div className={styled.divTitleTarefas}>
@@ -150,8 +161,8 @@ export const ModalEditCard = ({ infoCard, setOpenModalEdit }:{ infoCard: iCardSu
                         </div>
                         <div className={styled.divInput}>
                             <div className={styled.divAddTarefa}>
-                                <input type="text" placeholder="Digite a nova tarefa" value={novaTarefa} onChange={(e) => setNovaTarefa(e.target.value)}/>
-                                <button type="button" onClick={handleCriarTarefa}>Criar</button>
+                                <input type="text" placeholder="Digite a nova tarefa" value={novaTarefa} onChange={(e) => setNovaTarefa(e.target.value)} disabled={!authorized}/>
+                                <button type="button" onClick={handleCriarTarefa} disabled={!authorized}>Criar</button>
                             </div>
                             <div className={styled.divUl}>
                                 <ul className={styled.ul}>
@@ -159,11 +170,11 @@ export const ModalEditCard = ({ infoCard, setOpenModalEdit }:{ infoCard: iCardSu
                                         <li className={styled.li} key={index}>
                                             <div>
                                                 <div>
-                                                <input type="checkbox" checked={tarefa.completed} onChange={() => handleCheckboxChange(index)}/> 
-                                                 <label className={styled.newLabol} htmlFor={`tarefa-${index}`}>{tarefa.task}</label>
+                                                    <input type="checkbox" checked={tarefa.completed} onChange={() => handleCheckboxChange(index)} disabled={!authorized}/> 
+                                                    <label className={styled.newLabol} htmlFor={`tarefa-${index}`}>{tarefa.task}</label>
                                                 </div>
                                             </div>
-                                            <button type="button" onClick={() => excluirTarefa(index, tarefa.id)}>X</button>
+                                            <button type="button" onClick={() => excluirTarefa(index, tarefa.id)} disabled={!authorized}>X</button>
                                         </li>
                                     ))}
                                 </ul>
@@ -172,23 +183,26 @@ export const ModalEditCard = ({ infoCard, setOpenModalEdit }:{ infoCard: iCardSu
                     </div>
                     <div className={styled.divDesc}>
                         <p className={styled.pDesc}>Solução</p>
-                        <textarea className={styled.textarea} id="solution" {...register("solution")} defaultValue={infoCard.solution || ''}></textarea>
+                        <textarea className={styled.textarea} id="solution" {...register("solution")} defaultValue={infoCard.solution || ''} readOnly={!authorized}></textarea>
                     </div>
-                    <div className={styled.divExcluir}>
-                        <button type='submit' className={styled.salvar}>Salvar</button>
-                        <button type='button' onClick={() => setConfirmacaoExclusao(true)} className={styled.excluir}>Excluir</button>
-                        {confirmacaoExclusao && (
-                            <span className={styled.spanExcluir}>
-                                <p className={styled.pExcluir}>Tem certeza que deseja Excluir?</p>
-                                <div className={styled.divSimNao}>
-                                    <button className={styled.btnSim} type='button' onClick={() => confirmaExcluir(infoCard.id)}>Excluir</button>
-                                    <button className={styled.btnNao} type='button' onClick={() => setConfirmacaoExclusao(false)}>Não Excluir</button>
-                                </div>
-                            </span>
-                        )}
-                    </div>
+                    {authorized ? (
+                        <div className={styled.divExcluir}>
+                            <button type='submit' className={styled.salvar}>Salvar</button>
+                            <button type='button' onClick={() => setConfirmacaoExclusao(true)} className={styled.excluir}>Excluir</button>
+                            {confirmacaoExclusao && (
+                                <span className={styled.spanExcluir}>
+                                    <p className={styled.pExcluir}>Tem certeza que deseja Excluir?</p>
+                                    <div className={styled.divSimNao}>
+                                        <button className={styled.btnSim} type='button' onClick={() => confirmaExcluir(infoCard.id)}>Excluir</button>
+                                        <button className={styled.btnNao} type='button' onClick={() => setConfirmacaoExclusao(false)}>Não Excluir</button>
+                                    </div>
+                                </span>
+                            )}
+                        </div>
+                    ) : null }
                 </form>
             </div>
         </section>
     );
+    
 };
